[CELEBORN-78][REFACTOR] Extract heartbearter from LifecycleManager (#1021)
* [CELEBORN-78][REFACTOR] Extract heartbearter from LifecycleManager
